fix(load): store ptrMPK without truncating to int

+++ game/LOAD/LOAD_Callbacks.c
@@ -89,7 +89,7 @@ void LOAD_Callback_PatchMem(struct LoadQueueSlot *lqs)
 void LOAD_Callback_DriverModels(struct LoadQueueSlot *lqs)
 {
 	sdata->load_inProgress = 0;
-	sdata->ptrMPK = (int)lqs->ptrDestination;
+	sdata->ptrMPK = lqs->ptrDestination;
 }
 
 // NOTE(aalhendi): ASM-verified NTSC-U 926 0x80031b14-0x80031b50.
